Add common record, charge dataclasses.
These encompass the minimal fields we wish to collect and may be
subclassed to add additional fields if that's desired. Since the
FL/Bay/Court scraper has been populating additional fields for a while,
have it use a subclassed BenchmarkRecord (with accompanying builder) to
collect its scrapings.
